backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
pandas>=2.0
openpyxl>=3.1
numpy>=1.24
python-dotenv>=1.0
django-cors-headers>=4.0
openai>=1.0